src/src/test/java/GestionTurnoServiceTest.java
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.miapp.Clases.Medicamento;
import org.miapp.Clases.Medico;
import org.miapp.Clases.Paciente;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.mockito.junit.MockitoJUnitRunner;
import org.miapp.Servicios.GestionTurnoService;
import org.miapp.DAO.TurnoDAO;
import org.miapp.Clases.Turno;

import java.util.Collections;
import java.util.List;

import static org.junit.Assert.*;
import static org.mockito.Mockito.*;

@RunWith(MockitoJUnitRunner.class)
public class GestionTurnoServiceTest {

    @Mock
    private TurnoDAO turnoDAO;

    @InjectMocks
    private GestionTurnoService gestionTurnoService;

    @Mock
    private Turno turno;
    
    @Before
    public void setUp() {
        MockitoAnnotations.openMocks(this);
        // Configura el comportamiento del método obtenerTurnoPorId para devolver un turno cuando se le pase un ID válido
        Turno turnoEjemplo = new Turno(1, "Paciente Ejemplo", Collections.singletonList(new Medico("Médico Ejemplo")).toString(), true, false, String.valueOf(Collections.emptyList()));
        when(turnoDAO.obtenerTurnoPorId(anyInt())).thenReturn(turnoEjemplo);

        // Configura el comportamiento del método obtenerTurnoPorId para devolver null cuando se le pase un ID inválido
        when(turnoDAO.obtenerTurnoPorId(-1)).thenReturn(null);
        // Configura el comportamiento del método obtenerTurnoPorId para devolver null cuando se le pase un ID inválido
        // Esto simula que el turno no existe sin lanzar una excepción
        when(turnoDAO.obtenerTurnoPorId(-1)).thenReturn(null);

        // Configura el comportamiento del método crearTurno para devolver un turno específico después de ser creado
        when(turnoDAO.crearTurno(any(Turno.class))).thenReturn(new Turno(1, "Paciente Ejemplo", "Médico Ejemplo", true, false, "Receta Ejemplo"));

        // Configura el comportamiento del método actualizarTurno para devolver el turno actualizado
        when(turnoDAO.actualizarTurno(any(Turno.class))).thenAnswer(invocation -> {
            Turno turno = invocation.getArgument(0);
            // Simula la lógica de actualización aquí
            return turno;
        });
    }

    @Test
    public void testObtenerTurnoPorId() {
        // Llama al método que estás probando
        Turno result = gestionTurnoService.obtenerTurnoPorId(1);

        // Verifica que el resultado no sea nulo
        assertNotNull(result);
        // Verifica que el resultado tenga los valores esperados (puedes ajustar estos valores según tu lógica)
        assertEquals(1, result.getId());
        assertEquals("Paciente Ejemplo", result.getPaciente());
        assertEquals(1, result.getMedicosAsignados().size());
        assertEquals("Médico Ejemplo", result.getMedicosAsignados().get(0).getNombre());
        assertTrue(result.isEstaIniciado());
        assertFalse(result.isEstaCompletado());
        assertTrue(result.getReceta().isEmpty());

        // Verifica que el mock fue llamado con el ID correcto
        verify(turnoDAO).obtenerTurnoPorId(1);
    }

    @Test
    public void testAgregarTurno() {
        // Crear un turno para la prueba
        Turno turno = new Turno(1, "Paciente Ejemplo", "Médico Ejemplo", true, false, "Receta Ejemplo");

        // Configurar el comportamiento simulado del DAO
        when(turnoDAO.crearTurno(any(Turno.class))).thenReturn(turno);

        // Ejecutar el método que quieres probar
        Turno resultado = gestionTurnoService.agregarTurno(turno);

        // Verificar que el resultado no sea nulo
        assertNotNull(resultado);
        // Verificar que el resultado sea igual al turno creado
        assertEquals(turno, resultado);
    }

    @Test
    public void testEliminarTurnoNoExistente() {
        // Configurar el comportamiento simulado del DAO para lanzar una excepción cuando el turno no existe
        when(turnoDAO.obtenerTurnoPorId(999)).thenThrow(new IllegalArgumentException("El turno no existe"));

        // Ejecutar el método que quieres probar
        assertThrows(IllegalArgumentException.class, () -> gestionTurnoService.eliminarTurno(999));
    }

    @Test
        public void testActualizarTurno() {
            // Preparar datos de prueba
            Turno turnoActualizado = new Turno(1, "Paciente Ejemplo", "Médico Ejemplo", true, false, "Receta Ejemplo");
            turnoActualizado.setId(1); // Asume que este ID corresponde a un turno existente en tu simulación
            // Inicializa otros campos de turnoActualizado según sea necesario

            // Configurar el comportamiento simulado del DAO para un turno existente
            when(turnoDAO.actualizarTurno(any(Turno.class))).thenAnswer(invocation -> {
                // Simula la lógica de actualización aquí
                return invocation.<Turno>getArgument(0);
            });

            // Ejecutar el método que quieres probar
            Turno turnoResultante = turnoDAO.actualizarTurno(turnoActualizado);

            // Verificar que el turno resultante es el mismo que el turno actualizado
            assertEquals(turnoActualizado, turnoResultante);
        }


}

